refactor: moved internal packages of `arc` to `internal/`
