+++ arc/internal/compress/platform/platform.go
@@ -1,0 +1,10 @@
+package platform
+
+import (
+	"os"
+	"time"
+)
+
+func Timestamp(info os.FileInfo) (atime time.Time, mtime time.Time) {
+	return amTimes(info)
+}
