+++ arc/internal/compress/platform/read_darwin.go
@@ -1,7 +1,7 @@
-//go:build windows
-// +build windows
+//go:build darwin
+// +build darwin
 
-package arc
+package platform
 
 import (
 	"os"
@@ -11,9 +11,9 @@ import (
 
 // Возвращает временные метки доступа и изменения
 func amTimes(info os.FileInfo) (atime time.Time, mtime time.Time) {
-	stat := info.Sys().(*syscall.Win32FileAttributeData)
-	atime = time.Unix(0, stat.LastAccessTime.Nanoseconds())
-	mtime = time.Unix(0, stat.LastWriteTime.Nanoseconds())
+	stat := info.Sys().(*syscall.Stat_t)
+	atime = time.Unix(stat.Atimespec.Sec, stat.Atimespec.Nsec)
+	mtime = time.Unix(stat.Mtimespec.Sec, stat.Mtimespec.Nsec)
 
 	return atime, mtime
 }
