+++ arc/internal/generic/err.go
@@ -1,0 +1,5 @@
+package generic
+
+import "fmt"
+
+var ErrFlushWrBuf = fmt.Errorf("ошибка сброса буфера записи на диск")
