+++ arc/internal/header/err.go
