+++ arc/internal/header/symlink.go
